requirements.txt
praw>=7.7.0
google-generativeai>=0.8.0
pandas>=2.0.0
tqdm>=4.66.0 